+++ src/texture.zig
@@ -13,7 +13,7 @@ pub const SolidColor = struct {
         return .{ .albedo = albedo };
     }
 
-    pub fn value(self: SolidColor, u: f32, v: f32, p: Vec3) Color {
+    pub fn value(self: SolidColor, u: f64, v: f64, p: Vec3) Color {
         _ = u;
         _ = v;
         _ = p;
@@ -22,15 +22,15 @@ pub const SolidColor = struct {
 };
 
 pub const Checker = struct {
-    inv_scale: f32,
+    inv_scale: f64,
     even: *const Texture,
     odd: *const Texture,
 
-    pub fn init(scale: f32, even: *const Texture, odd: *const Texture) Checker {
+    pub fn init(scale: f64, even: *const Texture, odd: *const Texture) Checker {
         return .{ .inv_scale = 1.0 / scale, .even = even, .odd = odd };
     }
 
-    pub fn value(self: Checker, u: f32, v: f32, p: Vec3) Color {
+    pub fn value(self: Checker, u: f64, v: f64, p: Vec3) Color {
         const x_int: i32 = @intFromFloat(@floor(self.inv_scale * p.x));
         const y_int: i32 = @intFromFloat(@floor(self.inv_scale * p.y));
         const z_int: i32 = @intFromFloat(@floor(self.inv_scale * p.z));
@@ -46,7 +46,7 @@ pub const ImageTexture = struct {
         return .{ .image = image };
     }
 
-    pub fn value(self: ImageTexture, u: f32, v: f32, p: Vec3) Color {
+    pub fn value(self: ImageTexture, u: f64, v: f64, p: Vec3) Color {
         _ = p;
         if (self.image.height <= 0) return Color.init(0, 1, 1);
 
@@ -54,30 +54,30 @@ pub const ImageTexture = struct {
         const u_clamped = unit.clamp(u);
         const v_clamped = 1.0 - unit.clamp(v);
 
-        const w_f: f32 = @floatFromInt(self.image.width);
-        const h_f: f32 = @floatFromInt(self.image.height);
+        const w_f: f64 = @floatFromInt(self.image.width);
+        const h_f: f64 = @floatFromInt(self.image.height);
         const i: i32 = @intFromFloat(u_clamped * w_f);
         const j: i32 = @intFromFloat(v_clamped * h_f);
         const px = self.image.pixel(i, j);
 
         const scale_byte = 1.0 / 255.0;
         return Color.init(
-            @as(f32, @floatFromInt(px[0])) * scale_byte,
-            @as(f32, @floatFromInt(px[1])) * scale_byte,
-            @as(f32, @floatFromInt(px[2])) * scale_byte,
+            @as(f64, @floatFromInt(px[0])) * scale_byte,
+            @as(f64, @floatFromInt(px[1])) * scale_byte,
+            @as(f64, @floatFromInt(px[2])) * scale_byte,
         );
     }
 };
 
 pub const NoiseTexture = struct {
     noise: *const Perlin,
-    scale: f32,
+    scale: f64,
 
-    pub fn init(noise: *const Perlin, scale: f32) NoiseTexture {
+    pub fn init(noise: *const Perlin, scale: f64) NoiseTexture {
         return .{ .noise = noise, .scale = scale };
     }
 
-    pub fn value(self: NoiseTexture, u: f32, v: f32, p: Vec3) Color {
+    pub fn value(self: NoiseTexture, u: f64, v: f64, p: Vec3) Color {
         _ = u;
         _ = v;
         const phase = self.scale * p.z + 10.0 * self.noise.turb(p, 7);
@@ -95,7 +95,7 @@ pub const Texture = union(enum) {
         return .{ .solid = SolidColor.init(albedo) };
     }
 
-    pub fn value(self: Texture, u: f32, v: f32, p: Vec3) Color {
+    pub fn value(self: Texture, u: f64, v: f64, p: Vec3) Color {
         return switch (self) {
             inline else => |variant| variant.value(u, v, p),
         };
